server/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
requests
opencc-python-reimplemented
google-generativeai
python-dotenv